+++ GamingManagement-service/src/main/java/cz/fi/muni/pa165/yellow_yak/facade/PlayerFacadeImpl.java
@@ -36,8 +36,11 @@ public class PlayerFacadeImpl implements PlayerFacade {
 
     @Override
     public PlayerDTO create(String name, String email, Long teamId) {
-        if (name == null || email == null) {
+        if (name == null || email == null || teamId == null) {
             throw new IllegalArgumentException("arguments cannot be null");
+        }
+        if (teamId <= 0 || name.isEmpty() || email.isEmpty()) {
+            throw new IllegalArgumentException("invalid argument");
         }
         log.info("creating player, name = {}, email = {}, teamId = {}", name, email, teamId);
         return beanMappingService.mapTo(playerService.create(name, email, teamId), PlayerDTO.class);
@@ -48,6 +51,9 @@ public class PlayerFacadeImpl implements PlayerFacade {
         if (id == null) {
             throw new IllegalArgumentException("arguments cannot be null");
         }
+        if (id <= 0) {
+            throw new IllegalArgumentException("id cannot be zero or negative value");
+        }
         log.info("removing player, id = {}", id);
         return playerService.remove(id);
     }
@@ -56,6 +62,9 @@ public class PlayerFacadeImpl implements PlayerFacade {
     public PlayerDTO findById(Long id) {
         if (id == null) {
             throw new IllegalArgumentException("arguments cannot be null");
+        }
+        if (id <= 0) {
+            throw new IllegalArgumentException("id cannot be zero or negative value");
         }
         log.info("finding player by ID, id = {}", id);
         return beanMappingService.mapTo(playerService.findById(id), PlayerDTO.class);
@@ -66,6 +75,9 @@ public class PlayerFacadeImpl implements PlayerFacade {
         if (username == null) {
             throw new IllegalArgumentException("arguments cannot be null");
         }
+        if (username.isEmpty()) {
+            throw new IllegalArgumentException("invalid argument");
+        }
         log.info("listing player by username, username = {}", username);
         return beanMappingService.mapTo(playerService.findByUsername(username), PlayerDTO.class);
     }
@@ -74,6 +86,9 @@ public class PlayerFacadeImpl implements PlayerFacade {
     public List<PlayerDTO> findByTeam(Long teamId) {
         if (teamId == null) {
             throw new IllegalArgumentException("arguments cannot be null");
+        }
+        if (id <= 0) {
+            throw new IllegalArgumentException("teamId cannot be zero or negative value");
         }
         log.info("listing player by team, teamId = {}", teamId);
         return beanMappingService.mapTo(playerService.findByTeam(teamId), PlayerDTO.class);
